+++ xarray/plot/utils.py
@@ -3,28 +3,12 @@ from __future__ import absolute_import, division, print_function
 import warnings
 
 import numpy as np
-import pandas as pd
-import pkg_resources
 import textwrap
 
 from ..core.pycompat import basestring
 from ..core.utils import is_scalar
 
 ROBUST_PERCENTILE = 2.0
-
-
-def _load_default_cmap(fname='default_colormap.csv'):
-    """
-    Returns viridis color map
-    """
-    from matplotlib.colors import LinearSegmentedColormap
-
-    # Not sure what the first arg here should be
-    f = pkg_resources.resource_stream(__name__, fname)
-    cm_data = pd.read_csv(f, header=None).values
-    f.close()
-
-    return LinearSegmentedColormap.from_list('viridis', cm_data)
 
 
 def import_seaborn():
@@ -225,10 +209,6 @@ def _determine_cmap_params(plot_data, vmin=None, vmax=None, cmap=None,
             cmap = "RdBu_r"
         else:
             cmap = "viridis"
-
-    # Allow viridis before matplotlib 1.5
-    if cmap == "viridis":
-        cmap = _load_default_cmap()
 
     # Handle discrete levels
     if levels is not None: