+++ xarray/tests/test_dataset.py
@@ -2,10 +2,10 @@
 from __future__ import absolute_import, division, print_function
 
 from copy import copy, deepcopy
-from distutils.version import LooseVersion
 from io import StringIO
 from textwrap import dedent
 import warnings
+import sys
 
 import numpy as np
 import pandas as pd
@@ -94,15 +94,15 @@ class TestDataset(TestCase):
         <xarray.Dataset>
         Dimensions:  (dim1: 8, dim2: 9, dim3: 10, time: 20)
         Coordinates:
-          * time     (time) datetime64[ns] 2000-01-01 2000-01-02 2000-01-03 ...
+          * time     (time) datetime64[ns] 2000-01-01 2000-01-02 ... 2000-01-20
           * dim2     (dim2) float64 0.0 0.5 1.0 1.5 2.0 2.5 3.0 3.5 4.0
           * dim3     (dim3) %s 'a' 'b' 'c' 'd' 'e' 'f' 'g' 'h' 'i' 'j'
             numbers  (dim3) int64 0 1 2 0 0 1 1 2 2 3
         Dimensions without coordinates: dim1
         Data variables:
-            var1     (dim1, dim2) float64 -1.086 0.9973 0.283 -1.506 -0.5786 1.651 ...
-            var2     (dim1, dim2) float64 1.162 -1.097 -2.123 1.04 -0.4034 -0.126 ...
-            var3     (dim3, dim1) float64 0.5565 -0.2121 0.4563 1.545 -0.2397 0.1433 ...
+            var1     (dim1, dim2) float64 -1.086 0.9973 0.283 ... 0.1995 0.4684 -0.8312
+            var2     (dim1, dim2) float64 1.162 -1.097 -2.123 ... 0.1302 1.267 0.3328
+            var3     (dim3, dim1) float64 0.5565 -0.2121 0.4563 ... -0.2452 -0.3616
         Attributes:
             foo:      bar""") % data['dim3'].dtype  # noqa: E501
         actual = '\n'.join(x.rstrip() for x in repr(data).split('\n'))
@@ -183,15 +183,16 @@ class TestDataset(TestCase):
         data = Dataset({u'foø': [u'ba®']}, attrs={u'å': u'∑'})
         repr(data)  # should not raise
 
+        byteorder = '<' if sys.byteorder == 'little' else '>'
         expected = dedent(u"""\
         <xarray.Dataset>
         Dimensions:  (foø: 1)
         Coordinates:
-          * foø      (foø) <U3 %r
+          * foø      (foø) %cU3 %r
         Data variables:
             *empty*
         Attributes:
-            å:        ∑""" % u'ba®')
+            å:        ∑""" % (byteorder, u'ba®'))
         actual = unicode_type(data)
         assert expected == actual
 
@@ -846,7 +847,7 @@ class TestDataset(TestCase):
                     inds = np.nonzero(np.array(data[v].dims) == d)[0]
                     for ind in inds:
                         slice_list[ind] = s
-            expected = data[v].values[slice_list]
+            expected = data[v].values[tuple(slice_list)]
             actual = ret[v].values
             np.testing.assert_array_equal(expected, actual)
 
@@ -4036,9 +4037,6 @@ class TestDataset(TestCase):
             actual = ds.sortby(ds['A'])
         assert "DataArray is not 1-D" in str(excinfo.value)
 
-        if LooseVersion(np.__version__) < LooseVersion('1.11.0'):
-            pytest.skip('numpy 1.11.0 or later to support object data-type.')
-
         expected = sorted1d
         actual = ds.sortby('x')
         assert_equal(actual, expected)
@@ -4220,6 +4218,11 @@ def test_dataset_constructor_aligns_to_explicit_coords(
     result = xr.Dataset({'a': a}, coords=coords)
 
     assert_equal(expected, result)
+
+
+def test_error_message_on_set_supplied():
+    with pytest.raises(TypeError, message='has invalid type set'):
+        xr.Dataset(dict(date=[1, 2, 3], sec={4}))
 
 
 @pytest.mark.parametrize('unaligned_coords', (