+++ xarray/tests/test_plot.py
@@ -17,7 +17,8 @@ from xarray.plot.utils import (
 
 from . import (
     TestCase, assert_array_equal, assert_equal, raises_regex,
-    requires_matplotlib, requires_seaborn, requires_cftime)
+    requires_matplotlib, requires_matplotlib2, requires_seaborn,
+    requires_cftime)
 
 # import mpl and change the backend before other mpl imports
 try:
@@ -228,9 +229,31 @@ class TestPlot(PlotTestCase):
         np.testing.assert_allclose(xref, x)
         np.testing.assert_allclose(yref, y)
 
-        # test that warning is raised for non-monotonic inputs
+        # test that ValueError is raised for non-monotonic 1D inputs
         with pytest.raises(ValueError):
-            _infer_interval_breaks(np.array([0, 2, 1]))
+            _infer_interval_breaks(np.array([0, 2, 1]), check_monotonic=True)
+
+    def test_geo_data(self):
+        # Regression test for gh2250
+        # Realistic coordinates taken from the example dataset
+        lat = np.array([[16.28, 18.48, 19.58, 19.54, 18.35],
+                        [28.07, 30.52, 31.73, 31.68, 30.37],
+                        [39.65, 42.27, 43.56, 43.51, 42.11],
+                        [50.52, 53.22, 54.55, 54.50, 53.06]])
+        lon = np.array([[-126.13, -113.69, -100.92, -88.04, -75.29],
+                        [-129.27, -115.62, -101.54, -87.32, -73.26],
+                        [-133.10, -118.00, -102.31, -86.42, -70.76],
+                        [-137.85, -120.99, -103.28, -85.28, -67.62]])
+        data = np.sqrt(lon ** 2 + lat ** 2)
+        da = DataArray(data, dims=('y', 'x'),
+                       coords={'lon': (('y', 'x'), lon),
+                               'lat': (('y', 'x'), lat)})
+        da.plot(x='lon', y='lat')
+        ax = plt.gca()
+        assert ax.has_data()
+        da.plot(x='lat', y='lon')
+        ax = plt.gca()
+        assert ax.has_data()
 
     def test_datetime_dimension(self):
         nrow = 3
@@ -261,6 +284,7 @@ class TestPlot(PlotTestCase):
             d[0].plot(x='x', y='y', col='z', ax=plt.gca())
 
     @pytest.mark.slow
+    @requires_matplotlib2
     def test_subplot_kws(self):
         a = easy_array((10, 15, 4))
         d = DataArray(a, dims=['y', 'x', 'z'])
@@ -273,12 +297,9 @@ class TestPlot(PlotTestCase):
             cmap='cool',
             subplot_kws=dict(facecolor='r'))
         for ax in g.axes.flat:
-            try:
-                # mpl V2
-                assert ax.get_facecolor()[0:3] == \
-                    mpl.colors.to_rgb('r')
-            except AttributeError:
-                assert ax.get_axis_bgcolor() == 'r'
+            # mpl V2
+            assert ax.get_facecolor()[0:3] == \
+                mpl.colors.to_rgb('r')
 
     @pytest.mark.slow
     def test_plot_size(self):
@@ -422,10 +443,6 @@ class TestPlotHistogram(PlotTestCase):
         self.darray.plot.hist()
         assert 'testpoints [testunits]' == plt.gca().get_xlabel()
 
-    def test_ylabel_is_count(self):
-        self.darray.plot.hist()
-        assert 'Count' == plt.gca().get_ylabel()
-
     def test_title_is_histogram(self):
         self.darray.plot.hist()
         assert 'Histogram' == plt.gca().get_title()
@@ -461,7 +478,7 @@ class TestDetermineCmapParams(TestCase):
         cmap_params = _determine_cmap_params(self.data, robust=True)
         assert cmap_params['vmin'] == np.percentile(self.data, 2)
         assert cmap_params['vmax'] == np.percentile(self.data, 98)
-        assert cmap_params['cmap'].name == 'viridis'
+        assert cmap_params['cmap'] == 'viridis'
         assert cmap_params['extend'] == 'both'
         assert cmap_params['levels'] is None
         assert cmap_params['norm'] is None
@@ -545,7 +562,7 @@ class TestDetermineCmapParams(TestCase):
         cmap_params = _determine_cmap_params(pos)
         assert cmap_params['vmin'] == 0
         assert cmap_params['vmax'] == 1
-        assert cmap_params['cmap'].name == "viridis"
+        assert cmap_params['cmap'] == "viridis"
 
         # Default with negative data will be a divergent cmap
         cmap_params = _determine_cmap_params(neg)
@@ -557,17 +574,17 @@ class TestDetermineCmapParams(TestCase):
         cmap_params = _determine_cmap_params(neg, vmin=-0.1, center=False)
         assert cmap_params['vmin'] == -0.1
         assert cmap_params['vmax'] == 0.9
-        assert cmap_params['cmap'].name == "viridis"
+        assert cmap_params['cmap'] == "viridis"
         cmap_params = _determine_cmap_params(neg, vmax=0.5, center=False)
         assert cmap_params['vmin'] == -0.1
         assert cmap_params['vmax'] == 0.5
-        assert cmap_params['cmap'].name == "viridis"
+        assert cmap_params['cmap'] == "viridis"
 
         # Setting center=False too
         cmap_params = _determine_cmap_params(neg, center=False)
         assert cmap_params['vmin'] == -0.1
         assert cmap_params['vmax'] == 0.9
-        assert cmap_params['cmap'].name == "viridis"
+        assert cmap_params['cmap'] == "viridis"
 
         # However, I should still be able to set center and have a div cmap
         cmap_params = _determine_cmap_params(neg, center=0)
@@ -597,17 +614,17 @@ class TestDetermineCmapParams(TestCase):
         cmap_params = _determine_cmap_params(pos, vmin=0.1)
         assert cmap_params['vmin'] == 0.1
         assert cmap_params['vmax'] == 1
-        assert cmap_params['cmap'].name == "viridis"
+        assert cmap_params['cmap'] == "viridis"
         cmap_params = _determine_cmap_params(pos, vmax=0.5)
         assert cmap_params['vmin'] == 0
         assert cmap_params['vmax'] == 0.5
-        assert cmap_params['cmap'].name == "viridis"
+        assert cmap_params['cmap'] == "viridis"
 
         # If both vmin and vmax are provided, output is non-divergent
         cmap_params = _determine_cmap_params(neg, vmin=-0.2, vmax=0.6)
         assert cmap_params['vmin'] == -0.2
         assert cmap_params['vmax'] == 0.6
-        assert cmap_params['cmap'].name == "viridis"
+        assert cmap_params['cmap'] == "viridis"
 
 
 @requires_matplotlib
@@ -1556,12 +1573,26 @@ class TestFacetedLinePlots(PlotTestCase):
                                         range(3), ['A', 'B', 'C', 'C++']],
                                 name='Cornelius Ortega the 1st')
 
+        self.darray.hue.name = 'huename'
+        self.darray.hue.attrs['units'] = 'hunits'
+        self.darray.x.attrs['units'] = 'xunits'
+        self.darray.col.attrs['units'] = 'colunits'
+        self.darray.row.attrs['units'] = 'rowunits'
+
     def test_facetgrid_shape(self):
         g = self.darray.plot(row='row', col='col', hue='hue')
         assert g.axes.shape == (len(self.darray.row), len(self.darray.col))
 
         g = self.darray.plot(row='col', col='row', hue='hue')
         assert g.axes.shape == (len(self.darray.col), len(self.darray.row))
+
+    def test_unnamed_args(self):
+        g = self.darray.plot.line('o--', row='row', col='col', hue='hue')
+        lines = [q for q in g.axes.flat[0].get_children()
+                 if isinstance(q, mpl.lines.Line2D)]
+        # passing 'o--' as argument should set marker and linestyle
+        assert lines[0].get_marker() == 'o'
+        assert lines[0].get_linestyle() == '--'
 
     def test_default_labels(self):
         g = self.darray.plot(row='row', col='col', hue='hue')
@@ -1667,3 +1698,68 @@ def test_plot_cftime_data_error():
     data = DataArray(data, coords=[np.arange(5)], dims=['x'])
     with raises_regex(NotImplementedError, 'cftime.datetime'):
         data.plot()
+
+
+test_da_list = [DataArray(easy_array((10, ))),
+                DataArray(easy_array((10, 3))),
+                DataArray(easy_array((10, 3, 2)))]
+
+
+@requires_matplotlib
+class TestAxesKwargs(object):
+    @pytest.mark.parametrize('da', test_da_list)
+    @pytest.mark.parametrize('xincrease', [True, False])
+    def test_xincrease_kwarg(self, da, xincrease):
+        plt.clf()
+        da.plot(xincrease=xincrease)
+        assert plt.gca().xaxis_inverted() == (not xincrease)
+
+    @pytest.mark.parametrize('da', test_da_list)
+    @pytest.mark.parametrize('yincrease', [True, False])
+    def test_yincrease_kwarg(self, da, yincrease):
+        plt.clf()
+        da.plot(yincrease=yincrease)
+        assert plt.gca().yaxis_inverted() == (not yincrease)
+
+    @pytest.mark.parametrize('da', test_da_list)
+    @pytest.mark.parametrize('xscale', ['linear', 'log', 'logit', 'symlog'])
+    def test_xscale_kwarg(self, da, xscale):
+        plt.clf()
+        da.plot(xscale=xscale)
+        assert plt.gca().get_xscale() == xscale
+
+    @pytest.mark.parametrize('da', [DataArray(easy_array((10, ))),
+                                    DataArray(easy_array((10, 3)))])
+    @pytest.mark.parametrize('yscale', ['linear', 'log', 'logit', 'symlog'])
+    def test_yscale_kwarg(self, da, yscale):
+        plt.clf()
+        da.plot(yscale=yscale)
+        assert plt.gca().get_yscale() == yscale
+
+    @pytest.mark.parametrize('da', test_da_list)
+    def test_xlim_kwarg(self, da):
+        plt.clf()
+        expected = (0.0, 1000.0)
+        da.plot(xlim=[0, 1000])
+        assert plt.gca().get_xlim() == expected
+
+    @pytest.mark.parametrize('da', test_da_list)
+    def test_ylim_kwarg(self, da):
+        plt.clf()
+        da.plot(ylim=[0, 1000])
+        expected = (0.0, 1000.0)
+        assert plt.gca().get_ylim() == expected
+
+    @pytest.mark.parametrize('da', test_da_list)
+    def test_xticks_kwarg(self, da):
+        plt.clf()
+        da.plot(xticks=np.arange(5))
+        expected = np.arange(5).tolist()
+        assert np.all(plt.gca().get_xticks() == expected)
+
+    @pytest.mark.parametrize('da', test_da_list)
+    def test_yticks_kwarg(self, da):
+        plt.clf()
+        da.plot(yticks=np.arange(5))
+        expected = np.arange(5)
+        assert np.all(plt.gca().get_yticks() == expected)